distance attenuation in cpp to gain in csound
uses software channel to control control variable kamp for instrument 1. amplitude is found from user-to-rabbit distance

+++ rabbit.h
@@ -9,23 +9,24 @@ public:
    Rabbit();
    ~Rabbit();
 
-   // MODIFICATION MEMBER FUNCTIONS
+   // MODIFICATION MEMBER FUNCTIONS mutators
    void setTrueX(int);
    void setTrueY(int);
    void calcDistanceToUser(int, int);
    void activate(bool);
 
-   // CONSTANT MEMBER FUNCTIONS
+   // CONSTANT MEMBER FUNCTIONS accessors
    int getTrueX();
    int getTrueY();
-   int getDistanceToUser();
+   float getDistanceToUser();
    bool isActivated();
    Rabbit& getRabbit();
 
 private:
    int truex;
    int truey;
-   int distance_to_user;
+   ;int distance_to_user;
+   float distance_to_user;
    bool activated;
 	bool is_fox;
 };
